chore(team): split run lifecycle manager tests
